Create README.md
## EncryptStrings

Strings encryption library with dnlib. 

## Information

This library can be used to encrypt strings of .NET assemblies.
Base64 is currently used, but we can implement and add any encrypt and decrypt we need.

## Usage
Just download it and use in your .net applications. 

## Dependency  List of other open source libraries used by EncryptStrings
The EncryptStrings library has a dependency of dnlib (https://github.com/Xairooo/dnlib)